Removed references to 'browser globals'  from server-side code. https://www.gatsbyjs.org/docs/debugging-html-builds/

+++ src/components/Header/index.js
@@ -1,15 +1,19 @@
-import React, {useContext} from 'react'
-import {Responsive} from 'semantic-ui-react'
-import {navigate} from 'gatsby'
+import React, { useContext } from 'react';
+import { Responsive } from 'semantic-ui-react';
+import { navigate } from 'gatsby';
 
-import DesktopMenu from './DesktopMenu'
-import MobileMenu from './MobileMenu'
+import DesktopMenu from './DesktopMenu';
+import MobileMenu from './MobileMenu';
 
-const Header = ({location}) => {
-  const token = localStorage.getItem('jwt')
-  const signOut = () => {
-    localStorage.clear()
-    navigate('/')
+const Header = ({ location }) => {
+  let token;
+  let signOut;
+  if (typeof (localStorage) !== 'undefined') {
+    token = localStorage.getItem('jwt');
+    signOut = () => {
+      localStorage.clear();
+      navigate('/');
+    };
   }
 
   return (
@@ -21,7 +25,7 @@ const Header = ({location}) => {
         <DesktopMenu location={location} token={token} signout={signOut} />
       </Responsive>
     </>
-  )
-}
+  );
+};
 
-export default Header
+export default Header;
